requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
langchain-community>=0.0.20

# Web Scraping & API
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
google-api-python-client>=2.100.0
snscrape>=0.7.0.20230622

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Sentiment Analysis
textblob>=0.17.0
vaderSentiment>=3.3.2

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0

# PDF Generation
reportlab>=4.0.0
Pillow>=10.0.0
matplotlib>=3.7.0

# Twitter API
tweepy>=4.14.0

